runtimeConfig + refactor
