+++ src/features/index.ts
@@ -4,6 +4,7 @@ import memberUpdateLogger from './memberUpdateLogger'
 import nameChecker from './nameChecker'
 import ping from './ping'
 import preventEmojiSpam from './preventEmojiSpam'
+import runtimeConfig from './runtimeConfig'
 import stickyMessage from './stickyMessage'
 import threadPruner from './threadPruner'
 
@@ -12,6 +13,7 @@ export default [
   nameChecker,
   ping,
   preventEmojiSpam,
+  runtimeConfig,
   stickyMessage,
   threadPruner,
 ] as Plugin[]
