+++ src/features/runtimeConfig/index.ts
@@ -1,0 +1,12 @@
+import { definePlugin } from '@/types/definePlugin'
+
+import inspectConfig from './inspectConfig'
+import reloadConfig from './reloadConfig'
+
+export default definePlugin({
+  name: 'runtimeConfig',
+  setup: (pluginContext) => {
+    pluginContext.addCommand(inspectConfig)
+    pluginContext.addCommand(reloadConfig)
+  },
+})
